Move test jobs that are used for line number tests to separate file
